Override the system exit().  This is necessary to handle multiple calls to exit(), since the atexit handlers are only run once.  The claim in rev 49 of correctly handling this case now actually works, at least in the dynamic case, and the exit.c test program confirms this.  There is still one remaining step for the static case.
Testing for this problem revealed a race condition in the system
exit() between _IO_cleanup() and doing output from monitor's library
fini destructor that can (rarely) result in a segfault.  We delay the
secondary calls to exit() a few extra seconds, and for all practical
purposes, that avoids the problem.