Pull in the patch from hpctoolkit-externals.
1. Fix a bug where monitor failed to reinit the child correctly in the
case that a process forks and then creates a side thread all in an
init constructor before main.

2. Workaround a problem where side threads don't respond to signals.
In this case, we have to skip the fini-thread callback because calling
fini-thread after fini-process causes hpcrun to abort.

This is only a partial solution.  For the other half, we will need to
override sigwait() and friends.