Rework how monitor chooses the shootdown signal.  Keep track of the application's signal mask and don't let the application block all of our available signals.
Activate the --enable-open-signals option (maybe a bit iffy on the
name).  This allows the client to specify a list of signals at
configure time that the application is not allowed to block.

Change all of the locks to monitor's own spinlocks.  This way, we
never use pthread mutex locks.

Add the utils.c file for general purpose utilities.