Search all loaded objects on dlsym(RTLD_NEXT) fail
If a library is loaded with dlopen(RTLD_LOCAL), the libmonitor symbols
still override those provided by the loaded library, however libmonitor
is unable to find an implementation via dlsym(RTLD_NEXT) since the
object is not present in the global search list.

This commit adds a fallback when dlsym(RTLD_NEXT) fails, we now scan
through every loaded object and dlsym() it to find the symbol we want.
To prevent recursion with outer LD_PRELOAD wrappers, we start the search
only after encountering libmonitor itself. In short, this fallback is
similar to dlsym(RTLD_NEXT) but searches "deep" for the symbol, past the
standard search list.